@@ -16,7 +16,13 @@ public class RunnableTask implements officeTask, Runnable {
 
 	@Override
 	public void doTask() {
-
+		System.out.println(Thread.currentThread().getName() + " run do task");
+		try {
+			Thread.sleep(3000);
+		} catch (InterruptedException e) {
+			// TODO Auto-generated catch block
+			e.printStackTrace();
+		}
 	}
 
 }
